Now converts to XHTML using libtidy
libtidy provides the ability to convert HTML to XHTML. It is also able
to clean up the HTML a bit.
This unfortunately leads to the article DOM being serialized and
deserialized three times (Go after downloading, C to parse, then again
in Go), and the complete article existing in more than one place in
memory at once.